SPDB-702: Avoid unnecessary blocking io
`avoid_unnecessary_blocking_io` is an option in rocksdb that if set to `true`, allows working threads
to avoid doing unnecessary and long-latency operations on the “hot path” and instead schedule
background jobs to do them. We have seen improvements in compaction latency when this option
was set to `true`, due to the fact that obsolete files were deleted in a separate background
job, instead of being deleted as part of the compaction job.

Signed-off-by: Amnon Hanuhov <AmnonSWE@gmail.com>