makefile: fix parallel `make check` failing on the first run (#177)
After the changes in #141, parallel `make check` started failing with the
following GNU Parallel error:

> Option tmpdir requires an argument

This is caused by the fact that `test_config.mk` uses `?=` for setting
`TEST_TMPDIR` and on make restarts this doesn't work for some reason
(perhaps having to do with empty `TEST_TMPDIR` being somehow exported).

Replace it with a check for empty value and unconditional assignment
instead. While at it, include `test_config.mk` only on make restarts for
check targets instead of forcing everything that makes use of `TEST_TMPDIR`
to reuse the unit tests configuration.

Also use 6 Xs for the mktemp template in order to support BusyBox mktemp
and maybe some others as well, that require the template to end with 6 Xs
exactly.